refactor: decompose large modules + de-dup markdown/hydration (wave 3)
- store: extract restorePersistedTheme/restorePersistedUiPrefs shared by
  both hydration paths (removes ~40 duplicated lines)
- notes: extract PDF-export HTML prep into note-pdf-export.ts
- notes: extract shared ReactMarkdown pre→CodeBlock/Mermaid handler into
  markdown-code-fence.ts; used by note-editor, NoteMarkdownPreview, chat
- kanban: extract inline archive view into ArchiveView (board.tsx 657→549)

Slice split of terminal-sessions.ts skipped (high risk, no user benefit)
and GitView decomposition deferred. type-check:all + 836 tests + compile
all green.
