+++ src/components/kanban/board.tsx
@@ -26,7 +26,7 @@ import { useCairnStore } from "@/store";
 import { useShallow } from "zustand/react/shallow";
 import { Button } from "@/components/ui/button";
 import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogClose } from "@/components/ui/dialog";
-import { NoteMarkdownPreview } from "@/components/notes/NoteMarkdownPreview";
+import { ArchiveView } from "./archive-view";
 import { KanbanColumn } from "./column";
 import { KanbanCard } from "./card";
 import { CardDetailModal } from "./card-detail";
@@ -414,122 +414,14 @@ export function KanbanBoard() {
             </div>
           )}
           {/* ── Archive view ── */}
-          {archiveViewOpen && activeProjectId && (() => {
-            const allArchived = getArchivedProjectCards(activeProjectId);
-            const q = archiveFilter.toLowerCase();
-            const filtered = q
-              ? allArchived.filter((c) =>
-                  c.title.toLowerCase().includes(q) || (c.description ?? "").toLowerCase().includes(q)
-                )
-              : allArchived;
-            // Group by column name for display
-            const colMap = new Map(columns.map((c) => [c.id, c]));
-            return (
-              <div className="flex-1 flex flex-col min-h-0 overflow-hidden">
-                {/* Archive toolbar */}
-                <div className="flex items-center gap-2 px-4 py-2 border-b border-[var(--border)] flex-shrink-0">
-                  <div className="relative flex items-center flex-1 max-w-xs">
-                    <Search size={12} className="absolute left-2.5 text-[var(--text-tertiary)] pointer-events-none" />
-                    <input
-                      type="text"
-                      value={archiveFilter}
-                      onChange={(e) => setArchiveFilter(e.target.value)}
-                      placeholder="Search archived tasks…"
-                      className="w-full pl-7 pr-2 py-1.5 text-xs rounded-md bg-[var(--surface-2)] border border-[var(--border)] text-[var(--text-primary)] placeholder:text-[var(--text-tertiary)] focus:outline-none focus:border-[var(--accent)]"
-                    />
-                  </div>
-                  <span className="text-xs text-[var(--text-tertiary)]">
-                    {filtered.length} task{filtered.length !== 1 ? "s" : ""}
-                  </span>
-                  {filtered.length > 0 && (
-                    <Tooltip content={archiveFilter ? "Delete matching tasks permanently" : "Delete all archived tasks permanently"}>
-                      <button
-                        onClick={() => filtered.forEach((c) => deleteCard(c.id))}
-                        className="flex items-center gap-1.5 ml-auto px-2 py-1 rounded text-xs text-[var(--text-tertiary)] hover:text-[var(--danger)] hover:bg-[var(--danger)]/10 transition-colors"
-                      >
-                        <Trash2 size={11} />
-                        Delete all
-                      </button>
-                    </Tooltip>
-                  )}
-                </div>
-                {/* Archive card grid */}
-                <div className="flex-1 overflow-y-auto p-5">
-                  {filtered.length === 0 ? (
-                    <div className="flex flex-col items-center justify-center h-full gap-3">
-                      <ArchiveX size={32} className="text-[var(--text-tertiary)] opacity-30" />
-                      <p className="text-sm text-[var(--text-tertiary)]">
-                        {archiveFilter ? "No archived tasks match your search" : "No archived tasks"}
-                      </p>
-                    </div>
-                  ) : (
-                    <div className="grid grid-cols-[repeat(auto-fill,minmax(260px,1fr))] gap-3">
-                      {filtered.map((card) => {
-                        const col = colMap.get(card.columnId);
-                        return (
-                          <div key={card.id} className="rounded-xl border border-[var(--border)] bg-[var(--surface)] flex flex-col opacity-80 hover:opacity-100 transition-opacity">
-                            {/* Clickable body — opens detail modal */}
-                            <button
-                              onClick={() => setDetailCardId(card.id)}
-                              className="flex-1 p-3 text-left flex flex-col gap-2 hover:bg-[var(--surface-2)] rounded-t-xl transition-colors"
-                            >
-                              <span className="text-xs font-medium text-[var(--text-primary)] leading-snug line-clamp-2">{card.title}</span>
-                              {card.description && (
-                                <div className="text-[0.714rem] text-[var(--text-tertiary)] line-clamp-2">
-                                  <NoteMarkdownPreview content={card.description} className="!px-0 !py-0" />
-                                </div>
-                              )}
-                              <div className="flex items-center gap-1.5 mt-auto pt-1">
-                                {col && (
-                                  <span className="text-[0.643rem] px-1.5 py-0.5 rounded bg-[var(--surface-2)] text-[var(--text-tertiary)]">
-                                    {col.name}
-                                  </span>
-                                )}
-                                {card.priority && card.priority !== "medium" && (
-                                  <span className={cn(
-                                    "text-[0.643rem] px-1.5 py-0.5 rounded",
-                                    card.priority === "urgent" && "bg-[var(--danger)]/10 text-[var(--danger)]",
-                                    card.priority === "high"   && "bg-[var(--warning)]/10 text-[var(--warning)]",
-                                    card.priority === "low"    && "bg-[var(--surface-2)] text-[var(--text-tertiary)]",
-                                  )}>
-                                    {card.priority}
-                                  </span>
-                                )}
-                                {card.archivedAt && (
-                                  <span className="text-[0.643rem] text-[var(--text-tertiary)] ml-auto">
-                                    {new Date(card.archivedAt).toLocaleDateString()}
-                                  </span>
-                                )}
-                              </div>
-                            </button>
-                            {/* Action row */}
-                            <div className="flex items-center justify-end gap-0.5 px-2 py-1.5 border-t border-[var(--border)]">
-                              <Tooltip content="Restore to board">
-                                <button
-                                  onClick={() => restoreCard(card.id)}
-                                  className="p-1 rounded hover:bg-[var(--accent)]/10 text-[var(--text-tertiary)] hover:text-[var(--accent)] transition-colors"
-                                >
-                                  <ArchiveRestore size={12} />
-                                </button>
-                              </Tooltip>
-                              <Tooltip content="Delete permanently">
-                                <button
-                                  onClick={() => deleteCard(card.id)}
-                                  className="p-1 rounded hover:bg-[var(--danger)]/10 text-[var(--text-tertiary)] hover:text-[var(--danger)] transition-colors"
-                                >
-                                  <Trash2 size={12} />
-                                </button>
-                              </Tooltip>
-                            </div>
-                          </div>
-                        );
-                      })}
-                    </div>
-                  )}
-                </div>
-              </div>
-            );
-          })()}
+          {archiveViewOpen && activeProjectId && (
+            <ArchiveView
+              projectId={activeProjectId}
+              filter={archiveFilter}
+              onFilterChange={setArchiveFilter}
+              onOpenCard={setDetailCardId}
+            />
+          )}
 
           {/* ── Kanban board ── */}
           {!archiveViewOpen && <SortableContext items={columns.map((c) => c.id)} strategy={horizontalListSortingStrategy}>
